requirements.txt
Django==1.7.4
Pillow==2.7.0
PyMySQL==0.6.3
argparse==1.2.1
django-extensions==1.5.0
django-tastypie==0.12.1
python-dateutil==2.4.0
python-mimeparse==0.1.4
pytz==2014.10
six==1.9.0
wsgiref==0.1.2
